Small code shake-up
Use rust modules to reflect the following code design:
- src/main.rs	cli, setup, logs, IO
- src/lib.rs	core logic (can be re-exported)
